+++ components/rsvpFormClient.tsx
@@ -101,7 +101,7 @@ export default function RsvpFormClient({
       });
       const data = await res.json();
       if (data.success) {
-        router.push(`/your-invitation/${data.partyID}`);
+        router.push(`/guest-information/${data.partyID}`);
         if (typeof window !== "undefined") {
           window.setTimeout(() => {
             window.scrollTo({ top: 0, behavior: "smooth" });
