+++ playwright-tests/02_invite_code.spec.ts
@@ -14,7 +14,7 @@ test("shows the not-found error when the invite code is rejected", async ({
     }),
   );
 
-  await page.goto("/your-invitation");
+  await page.goto("/guest-information");
 
   await page
     .getByRole("textbox", { name: /enter invite code/i })
@@ -24,7 +24,7 @@ test("shows the not-found error when the invite code is rejected", async ({
   await expect(page.getByText(/invite code not found/i)).toBeVisible();
 });
 
-test("navigates to /your-invitation/<code> on success", async ({ page }) => {
+test("navigates to /guest-information/<code> on success", async ({ page }) => {
   await page.route("**/api/checkInvite", (route) =>
     route.fulfill({
       status: 200,
@@ -35,7 +35,7 @@ test("navigates to /your-invitation/<code> on success", async ({ page }) => {
 
   // The dynamic [id] page does a server-side DB query we cannot intercept;
   // short-circuit it with a stub response so the spec stays self-contained.
-  await page.route("**/your-invitation/abc123", (route) =>
+  await page.route("**/guest-information/abc123", (route) =>
     route.fulfill({
       status: 200,
       contentType: "text/html",
@@ -43,13 +43,13 @@ test("navigates to /your-invitation/<code> on success", async ({ page }) => {
     }),
   );
 
-  await page.goto("/your-invitation");
+  await page.goto("/guest-information");
 
   await page
     .getByRole("textbox", { name: /enter invite code/i })
     .fill("ABC123");
   await page.getByRole("button", { name: /^submit$/i }).click();
 
-  await page.waitForURL("**/your-invitation/abc123");
-  expect(page.url()).toContain("/your-invitation/abc123");
+  await page.waitForURL("**/guest-information/abc123");
+  expect(page.url()).toContain("/guest-information/abc123");
 });
